746: move file handling into composable
