+++ frontend/components/modal/ModalUploadImage.vue
@@ -43,40 +43,8 @@
 
 <script setup lang="ts">
 import { DialogTitle } from '@headlessui/vue';
-import { ref } from 'vue';
 import ModalBase from '~/components/modal/ModalBase.vue';
+import useFileManager from '~/composables/useFileManager';
 
-const files = ref<UploadableFile[]>([]);
-  function handleFiles(newFiles: File[]) {
-        const newUploadableFiles = [...newFiles]
-            .map((file) => new UploadableFile(file))
-            .filter((file) => !fileExists(file.id))
-        files.value = [...files.value, ...newUploadableFiles];
-  }
-
-  function fileExists(otherId: string) {
-      return files.value.some(({ id }) => id === otherId)
-  }
-
-  function removeFile(file: UploadableFile) {
-      const index = files.value.indexOf(file)
-
-      if (index > -1) files.value.splice(index, 1)
-  }
-
-
-class UploadableFile {
-file: File;
-url: string;
-name: string;
-status: null;
-id: string;
-    constructor(file: File) {
-        this.file = file
-        this.name = file.name
-        this.id = `${file.name}-${file.size}-${file.lastModified}-${file.type}`
-        this.url = URL.createObjectURL(file)
-        this.status = null
-    }
-}
+const { files, handleFiles, removeFile } = useFileManager();
 </script>
