+++ frontend/composables/useFileManager.ts
@@ -1,0 +1,40 @@
+export default function useFileManager() {
+  const files = ref<UploadableFile[]>([]);
+  function handleFiles(newFiles: File[]) {
+    const newUploadableFiles = [...newFiles]
+      .map((file) => new UploadableFile(file))
+      .filter((file) => !fileExists(file.id));
+    files.value = [...files.value, ...newUploadableFiles];
+  }
+
+  function fileExists(otherId: string) {
+    return files.value.some(({ id }) => id === otherId);
+  }
+
+  function removeFile(file: UploadableFile) {
+    const index = files.value.indexOf(file);
+
+    if (index > -1) files.value.splice(index, 1);
+  }
+
+  return {
+    files,
+    handleFiles,
+    removeFile,
+  };
+}
+
+class UploadableFile {
+  file: File;
+  url: string;
+  name: string;
+  status: null;
+  id: string;
+  constructor(file: File) {
+    this.file = file;
+    this.name = file.name;
+    this.id = `${file.name}-${file.size}-${file.lastModified}-${file.type}`;
+    this.url = URL.createObjectURL(file);
+    this.status = null;
+  }
+}
